refactor: extract conversion and help logic from model
Move conversion message handling and the help view into dedicated
files to reduce the complexity of model.go.

Introduce handleConvertMsg to encapsulate the conversion lifecycle
and update the main model to delegate these messages. Reorganize
the model struct to group related state fields.
