+++ converter.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"fmt"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -51,6 +52,77 @@ func convertFile(ctx context.Context, src string) tea.Cmd {
 		}
 		return convertDoneMsg{src, dest}
 	}
+}
+
+// handleConvertMsg handles all conversion lifecycle messages. Returns false if
+// the message was not a conversion message.
+func handleConvertMsg(m model, msg tea.Msg) (model, tea.Cmd, bool) {
+	switch msg := msg.(type) {
+	case execConvertMsg:
+		if !m.ffmpegAvailable {
+			m.statusMsg = "ffmpeg not available"
+			m.statusIsError = true
+			return m, nil, true
+		}
+		if len(msg.files) == 0 {
+			m.statusMsg = "No convertible files selected"
+			m.statusIsError = true
+			return m, nil, true
+		}
+		m.convertQueue = msg.files
+		m.convertIndex = 0
+		m.convertDone = 0
+		m.convertSkipped = 0
+		m.convertErrors = 0
+		ctx, cancel := context.WithCancel(context.Background())
+		m.convertCtx = ctx
+		m.convertCancel = cancel
+		m.statusMsg = fmt.Sprintf("Converting 1/%d...", len(m.convertQueue))
+		m.statusIsError = false
+		return m, convertFile(m.convertCtx, m.convertQueue[0]), true
+	case convertDoneMsg:
+		m.convertDone++
+		m, cmd := nextConvert(m)
+		return m, cmd, true
+	case convertSkippedMsg:
+		m.convertSkipped++
+		m, cmd := nextConvert(m)
+		return m, cmd, true
+	case convertErrMsg:
+		m.convertErrors++
+		m, cmd := nextConvert(m)
+		return m, cmd, true
+	}
+	return m, nil, false
+}
+
+func nextConvert(m model) (model, tea.Cmd) {
+	if m.convertCancelled {
+		m.convertCancel = nil
+		m.convertCtx = nil
+		m.convertCancelled = false
+		m.convertQueue = nil
+		var teaCmd tea.Cmd
+		m.browser, teaCmd = m.browser.changeDir(m.browser.dir)
+		return m, teaCmd
+	}
+	m.convertIndex++
+	if m.convertIndex < len(m.convertQueue) {
+		m.statusMsg = fmt.Sprintf("Converting %d/%d...", m.convertIndex+1, len(m.convertQueue))
+		m.statusIsError = false
+		return m, convertFile(m.convertCtx, m.convertQueue[m.convertIndex])
+	}
+	if m.convertCancel != nil {
+		m.convertCancel()
+		m.convertCancel = nil
+		m.convertCtx = nil
+	}
+	m.statusMsg = fmt.Sprintf("Conversion complete (%d converted, %d skipped, %d errors)",
+		m.convertDone, m.convertSkipped, m.convertErrors)
+	m.statusIsError = m.convertErrors > 0
+	var teaCmd tea.Cmd
+	m.browser, teaCmd = m.browser.changeDir(m.browser.dir)
+	return m, teaCmd
 }
 
 // buildConvertList collects convertible files from the given entries.
